chore(tests): extracted phpunit helper classes from test files
ref #13143
