+++ engine/tests/classes/Elgg/Helpers/Application/FooController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Elgg\Helpers\Application;
+
+use Elgg\Request;
+use Elgg\Http\OkResponse;
+
+/**
+ * @see Elgg\ApplicationUnitTest
+ */
+class FooController {
+	public function __invoke(Request $request) {
+		$response = new OkResponse($request->getParam('echo'));
+		return $response;
+	}
+}
