+++ engine/tests/classes/Elgg/Helpers/CustomUser.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Elgg\Helpers;
+
+/**
+ * @see \Elgg\Application\BootHandlerUnitTest
+ * @see \Elgg\Users\AccountsServiceUnitTest
+ */
+class CustomUser extends \ElggUser {
+	
+}
