+++ engine/tests/classes/Elgg/Helpers/Database/DatabaseTestObj.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Elgg\Helpers\Database;
+
+/**
+ * @see \Elgg\DatabaseUnitTest
+ */
+class DatabaseTestObj {
+	
+	public function __invoke() {
+		
+	}
+	
+}
