+++ engine/tests/classes/Elgg/Helpers/Di/DiContainerTestObject.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Elgg\Helpers\Di;
+
+/**
+ * @see \Elgg\Di\DiContainerUnitTest
+ */
+class DiContainerTestObject {
+	
+	public $di;
+	
+	public function __construct($di = null) {
+		$this->di = $di;
+	}
+	
+}
