+++ engine/tests/classes/Elgg/Helpers/Di/ServiceFacadeTestService.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Elgg\Helpers\Di;
+
+use Elgg\Di\ServiceFacade;
+
+/**
+ * @see Elgg\Di\ServiceFacadeTest
+ */
+class ServiceFacadeTestService {
+	use ServiceFacade;
+	
+	public static function name() {
+		return 'foo';
+	}
+	
+	public function greet($name) {
+		return "Hi, $name";
+	}
+}
