+++ engine/tests/classes/Elgg/Helpers/ElggExtenderTestObject.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Elgg\Helpers;
+
+/**
+ * @see \ElggExtenderUnitTest
+ */
+class ElggExtenderTestObject {
+	function __toString() {
+		return 'oh hey';
+	}
+}
