+++ engine/tests/classes/Elgg/Helpers/ElggObjectWithExposableAttributes.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Elgg\Helpers;
+
+/**
+ * @see \Elgg\Integration\ElggCoreObjectTest
+ */
+class ElggObjectWithExposableAttributes extends \ElggObject {
+	public function expose_attributes() {
+		return $this->attributes;
+	}
+}
