+++ engine/tests/classes/Elgg/Helpers/ElggUserWithExposableAttributes.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Elgg\Helpers;
+
+/**
+ * @see \Elgg\Integration\ElggCoreUserTest
+ */
+class ElggUserWithExposableAttributes extends \ElggUser {
+	public function expose_attributes() {
+		return $this->attributes;
+	}
+}
