+++ engine/tests/classes/Elgg/Helpers/HooksRegistrationServiceTestInvokable.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Elgg\Helpers;
+
+/**
+ * @see \Elgg\HooksRegistrationServiceUnitTest
+ */
+class HooksRegistrationServiceTestInvokable {
+	
+	const KLASS = __CLASS__;
+	
+	function __invoke() {
+		
+	}
+}
