+++ engine/tests/classes/Elgg/Helpers/MethodMatcherTestObject.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Elgg\Helpers;
+
+/**
+ * @see \Elgg\MethodMatcherUnitTest
+ */
+class MethodMatcherTestObject {
+	
+}
