+++ engine/tests/classes/Elgg/Helpers/PreloaderMock20140623.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Elgg\Helpers;
+
+/**
+ * @see \Elgg\EntityPreloaderUnitTest
+ */
+class PreloaderMock20140623 {
+	
+	function isCached($guid) {
+		return $guid < 100;
+	}
+	
+	function load($opts) {
+		
+	}
+}
