+++ engine/tests/phpunit/integration/Elgg/Actions/Admin/UpgradeTest.php
@@ -7,6 +7,7 @@ use Elgg\EntityNotFoundException;
 use Elgg\Http\OkResponse;
 use Elgg\Upgrade\AsynchronousUpgrade;
 use Elgg\Upgrade\Result;
+use Elgg\Helpers\Actions\Admin\UpgradeTestBatch;
 
 /**
  * @group UpgradeService
@@ -63,34 +64,5 @@ class UpgradeTest extends ActionResponseTestCase {
 		if (!_elgg_services()->upgrades->getPendingUpgrades()) {
 			$this->assertFalse(elgg_admin_notice_exists('pending_upgrades'));
 		}
-	}
-}
-
-class UpgradeTestBatch implements AsynchronousUpgrade {
-
-	protected $_version;
-
-	public function getVersion() {
-		if (!isset($this->_version)) {
-			$this->_version = date('Ymd') . rand(10, 99);
-		}
-
-		return $this->_version;
-	}
-
-	public function needsIncrementOffset() {
-		return true;
-	}
-
-	public function shouldBeSkipped() {
-		return false;
-	}
-
-	public function countItems() {
-		return 100;
-	}
-
-	public function run(Result $result, $offset) {
-		$result->addSuccesses(10);
 	}
 }
