+++ engine/tests/phpunit/unit/Elgg/Collections/CollectionsUnitTest.php
@@ -3,6 +3,7 @@
 namespace Elgg\Collections;
 
 use Elgg\UnitTestCase;
+use Elgg\Helpers\Collections\TestItem;
 
 /**
  * @group Collections
@@ -290,30 +291,5 @@ class CollectionsUnitTest extends UnitTestCase {
 	public function testConstructorThrowsWithInvalidClass() {
 		$this->expectException(\InvalidArgumentException::class);
 		new Collection([], MyClass::class);
-	}
-
-}
-
-class TestItem implements CollectionItemInterface {
-
-	public function __construct($id, $priority) {
-		$this->id = $id;
-		$this->priority = $priority;
-	}
-
-	/**
-	 * Get unique item identifier within a collection
-	 * @return string|int
-	 */
-	public function getID() {
-		return $this->id;
-	}
-
-	/**
-	 * Get priority (weight) of the item within a collection
-	 * @return int
-	 */
-	public function getPriority() {
-		return $this->priority;
 	}
 }
