+++ engine/tests/phpunit/unit/Elgg/DeprecationWrapperUnitTest.php
@@ -2,6 +2,9 @@
 
 namespace Elgg;
 
+use Elgg\Helpers\DeprecationWrapperTestObj1;
+use Elgg\Helpers\DeprecationWrapperTestObj2;
+
 /**
  * @group UnitTests
  */
@@ -101,46 +104,6 @@ class DeprecationWrapperUnitTest extends \Elgg\UnitTestCase {
 		$wrapper[0] = 'value'; $line = __LINE__;
 		$this->assertEquals('value', $obj->{'0'});
 		$this->assertEquals("$file:$line", $this->last_stack_line);
-	}
-
-}
-
-class DeprecationWrapperTestObj1 {
-
-	public $foo = 'foo';
-
-	public function foo() {
-		return 'foo';
-	}
-
-	public function __toString() {
-		return 'foo';
-	}
-
-}
-
-class DeprecationWrapperTestObj2 extends \ArrayObject {
-
-	public $data = array();
-
-	public function offsetSet($offset, $value) {
-		if (is_null($offset)) {
-			$this->data[] = $value;
-		} else {
-			$this->data[$offset] = $value;
-		}
-	}
-
-	public function offsetExists($offset) {
-		return array_key_exists($offset, $this->data);
-	}
-
-	public function offsetUnset($offset) {
-		unset($this->data[$offset]);
-	}
-
-	public function offsetGet($offset) {
-		return isset($this->data[$offset]) ? $this->data[$offset] : null;
 	}
 
 }
